mem0_api/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
mem0ai==1.0.5
pymilvus>=2.5.0,<2.6.0
